initial libnode
