+++ libnode/Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bytes = "1.7.2"
+openssl = "0.10.66"
+ph = { path = "../adapter/ph" }
+rand = "0.8.5"
 thrift = "0.17.0"
 tokio = { version = "1.40.0", features = ["full"] }
 tokio-util = { version = "0.7.12", features = ["rt"] }
