+++ libnode/src/lib.rs
@@ -1,3 +1,6 @@
-mod vss;
+pub mod vss;
+pub mod vsconn;
+mod vscli;
 mod vsapi;
-mod vstypes;
+
+
