--- node/src/vs/mod.rs
@@ -1,7 +1,0 @@
-mod vsconn;
-pub use vsconn::VSConn;
-pub use vsconn::VSOutput;
-
-mod vscli;
-pub mod vss;
-mod vstypes;
